+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/GitWorkingDirectory.java
@@ -14,6 +14,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.jcraft.jsch.Session;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -26,7 +28,11 @@ import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.PushCommand;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.lib.Repository;
+import org.eclipse.jgit.lib.StoredConfig;
 import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
+import org.eclipse.jgit.transport.JschConfigSessionFactory;
+import org.eclipse.jgit.transport.OpenSshConfig;
+import org.eclipse.jgit.transport.SshSessionFactory;
 
 /**
  * @author Michael Hashimoto
@@ -35,6 +41,15 @@ public class GitWorkingDirectory {
 
 	public GitWorkingDirectory(String workingDirectory)
 		throws GitAPIException, InterruptedException, IOException {
+
+		this("master", workingDirectory);
+	}
+
+	public GitWorkingDirectory(
+			String upstreamBranchName, String workingDirectory)
+		throws GitAPIException, InterruptedException, IOException {
+
+		_upstreamBranchName = upstreamBranchName;
 
 		_setWorkingDirectory(workingDirectory);
 
@@ -49,6 +64,9 @@ public class GitWorkingDirectory {
 		_repository = fileRepositoryBuilder.build();
 
 		_git = new Git(_repository);
+
+		_repositoryName = _getRepositoryName();
+		_repositoryUsername = _getRepositoryUsername();
 	}
 
 	public void checkoutBranch(String branchName)
@@ -139,6 +157,18 @@ public class GitWorkingDirectory {
 		return _repository.getBranch();
 	}
 
+	public String getRepositoryName() {
+		return _repositoryName;
+	}
+
+	public String getRepositoryUsername() {
+		return _repositoryUsername;
+	}
+
+	public String getUpstreamBranchName() {
+		return _upstreamBranchName;
+	}
+
 	public File getWorkingDirectory() {
 		return _workingDirectory;
 	}
@@ -179,6 +209,37 @@ public class GitWorkingDirectory {
 		addCommand.call();
 	}
 
+	private String _getRepositoryName() {
+		StoredConfig storedConfig = _repository.getConfig();
+
+		String remote = storedConfig.getString("remote", "upstream", "url");
+
+		int x = remote.indexOf("/") + 1;
+		int y = remote.indexOf(".git");
+
+		String repositoryName = remote.substring(x, y);
+
+		if (!_upstreamBranchName.contains("ee-") &&
+			!_upstreamBranchName.contains("-private")) {
+
+			repositoryName = repositoryName.replace("-ee", "");
+			repositoryName = repositoryName.replace("-private", "");
+		}
+
+		return repositoryName;
+	}
+
+	private String _getRepositoryUsername() {
+		StoredConfig storedConfig = _repository.getConfig();
+
+		String remote = storedConfig.getString("remote", "upstream", "url");
+
+		int x = remote.indexOf(":") + 1;
+		int y = remote.indexOf("/");
+
+		return remote.substring(x, y);
+	}
+
 	private void _setWorkingDirectory(String workingDirectory)
 		throws GitAPIException, IOException {
 
@@ -215,9 +276,28 @@ public class GitWorkingDirectory {
 		}
 	}
 
+	static {
+		JschConfigSessionFactory jschConfigSessionFactory =
+			new JschConfigSessionFactory() {
+
+				@Override
+				protected void configure(
+					OpenSshConfig.Host host, Session session) {
+
+					session.setConfig("StrictHostKeyChecking", "no");
+				}
+
+			};
+
+		SshSessionFactory.setInstance(jschConfigSessionFactory);
+	}
+
 	private final Git _git;
 	private File _gitDirectory;
 	private final Repository _repository;
+	private final String _repositoryName;
+	private final String _repositoryUsername;
+	private final String _upstreamBranchName;
 	private File _workingDirectory;
 
 }