+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/JavaSourceProcessor.java
@@ -600,7 +600,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 			processMessage(
 				fileName,
 				"The declared package '" + packagePath +
-					"' does not match the expected package");
+					"' does not match the expected package",
+				"package.markdown");
 
 			return;
 		}
@@ -1759,7 +1760,7 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 
 		while (matcher.find()) {
 			if (getLevel(matcher.group()) == 0) {
-				content = StringUtil.replaceFirst(
+				return StringUtil.replaceFirst(
 					content, "\n", "\n\n", matcher.start());
 			}
 		}
@@ -1783,7 +1784,18 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 				matcher.end());
 
 			if ((x != -1) && (x != y)) {
-				content = StringUtil.replaceFirst(content, "\n", "\n\n", x + 1);
+				return StringUtil.replaceFirst(content, "\n", "\n\n", x + 1);
+			}
+		}
+
+		matcher = _missingEmptyLinePattern3.matcher(content);
+
+		while (matcher.find()) {
+			if ((getLevel(matcher.group()) != 0) &&
+				(content.charAt(matcher.end()) != CharPool.NEW_LINE)) {
+
+				return StringUtil.replaceFirst(
+					content, "\n", "\n\n", matcher.end() - 1);
 			}
 		}
 
@@ -4767,6 +4779,8 @@ public class JavaSourceProcessor extends BaseSourceProcessor {
 		"(\t| = |return )new .*\\(.*\\) \\{\n\t+[^{\t]");
 	private final Pattern _missingEmptyLinePattern2 = Pattern.compile(
 		"(\n\t*)(public|private|protected) [^;]+? \\{");
+	private final Pattern _missingEmptyLinePattern3 = Pattern.compile(
+		"\n.*\\) \\{\n");
 	private final Map<String, String> _moduleFileContentsMap =
 		new ConcurrentHashMap<>();
 	private Map<String, String> _moduleFileNamesMap;